104 -fib

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,7 +5,7 @@
  */
 int main(void)
 {
-	unsigned long long int c, p, num;
+	unsigned long int c, p, num;
 	int n;
 
 	c = 1;
